refactor: move functions to utils module
